+++ ChangeLog.md
@@ -1,9 +1,11 @@
 # Changelog
 
+## 1.1.3 (2019-04-15)
+- Accept CRAM format for the input file
+
 ## 1.1.2 (2018-07-22)
 - Chenge the prefix of ReadLengthDist.csv and FragmentLengthDist.csv to *.tsv
 - Add --allchr options (for DROMPAplus)
-
 
 ## 1.1.1 (2018-04-25)
 - Improve to read BAM file faster